requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.0
h5py==3.14.0
hdf5storage==0.2.0
hf-xet==1.1.5
huggingface-hub==0.34.0
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.1
pillow==11.2.1
platformdirs==4.3.8
portalocker==3.2.0
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.30.0
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==0.9.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
Werkzeug==3.1.3
